+++ src/widgets/file_row.cpp
@@ -47,7 +47,7 @@ LogEntry::LogEntry() : Glib::ObjectBase("MtSyncLogEntry") {}
 Glib::RefPtr<LogEntry> LogEntry::create(
     const std::string& time, const std::string& state,
     const std::string& job_id, const std::string& job_type,
-    const std::string& contents)
+    const std::string& contents, const std::string& log_path)
 {
     auto obj = Glib::make_refptr_for_instance(new LogEntry());
     obj->property_time.set_value(time);
@@ -55,6 +55,7 @@ Glib::RefPtr<LogEntry> LogEntry::create(
     obj->property_job_id.set_value(job_id);
     obj->property_job_type.set_value(job_type);
     obj->property_contents.set_value(contents);
+    obj->property_log_path.set_value(log_path);
     return obj;
 }