+++ src/widgets/file_row.hpp
@@ -59,13 +59,15 @@ public:
         const std::string& state,
         const std::string& job_id,
         const std::string& job_type,
-        const std::string& contents);
+        const std::string& contents,
+        const std::string& log_path = "");
 
     Glib::Property<Glib::ustring> property_time    {*this, "time"};
     Glib::Property<Glib::ustring> property_state   {*this, "state"};
     Glib::Property<Glib::ustring> property_job_id  {*this, "job-id"};
     Glib::Property<Glib::ustring> property_job_type{*this, "job-type"};
     Glib::Property<Glib::ustring> property_contents{*this, "contents"};
+    Glib::Property<Glib::ustring> property_log_path{*this, "log-path"};
 
 protected:
     LogEntry();